Use common cdp_interface.rb for both Chrome and Firefox, allow for FIREFOX_*, CHROME_* or common OPAL_CDP_* environment variables
